% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.SIM.R
\name{download.SIM}
\alias{download.SIM}
\title{Download SIM}
\usage{
download.SIM(inicio, fim, UF = "all")
}
\arguments{
\item{inicio}{The year that start the files extract.}

\item{fim}{The year that finish the files extract. By default the last year. Can't be the current year.}

\item{UF}{The state acronym.}
}
\value{
A data frame with the filtered raw SIM data.
}
\description{
Download SIM
}
\examples{
ce <- download.SIM(inicio = 2022, fim = 2022,UF= "CE")
}
\seealso{
\code{\link{process.SIM} \link{panel.SIM}}
}
\author{
Luan Augusto, \email{luanguto87@gmail.com}
}
