% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.SIM.R
\name{panel.SIM}
\alias{panel.SIM}
\title{Panel SIM}
\usage{
panel.SIM(data)
}
\arguments{
\item{data}{A data frame from SINASC with the columns needed for the Power BI dashboard.}
}
\description{
Panel SIM
}
\examples{
fSIM <- panel.SIM(db_label)
}
