% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.SIM.R
\name{process.SIM}
\alias{process.SIM}
\title{Process SIM}
\usage{
process.SIM(data)
}
\arguments{
\item{data}{Raw SIM data generated by the function download.SIM with all columns.}
}
\description{
Process SIM
}
