% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.SIM.R
\name{download.SIM}
\alias{download.SIM}
\title{Download SIM}
\usage{
download.SIM(uf, periodo, dir = ".", filename = NULL)
}
\arguments{
\item{uf}{The state acronym.}

\item{periodo}{A vector with one or two value(s): Strings indicating the period at the start of the interval of the required publications in the format "YYYY".}

\item{dir}{A string indicating the output directory where the files will be stored (the default is the directory created within the User's Documents folder).}

\item{filename}{A string indicating the name of the file for the downloaded file (the default is constructed based on the input parameters from the directory of the user's Documents folder).}
}
\description{
Download SIM
}
